requirements.txt
google-generativeai>=0.3.0aiohttp==3.11.18

requests>=2.31.0protobuf==6.31.0

urllib3>=2.0.0python-dotenv==1.1.0

quart==0.20.0
quart_cors==0.8.0
Requests==2.32.3
